llm_cloud_service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
google-generativeai
httpx
python-multipart
